ralphsaunders.co.uk theme underway.
I've created the beginings of a theme after following the documentation[1][2].
A master page file - `master.html` - and an index file - `index.html` - which
inherits from the master file.

You can see the title of the index page is set by extending the title block of
the master page.

So far, so easy.

[1]: http://docs.getpelican.com/en/latest/themes.html#theming-pelican
[2]: http://docs.getpelican.com/en/latest/settings.html#themes